backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-groq
chromadb
sentence-transformers
youtube-transcript-api
python-dotenv
